remove components built for supporterPlusOnly
